==> UltraJaya/Ultramilk_Final/1a_Merge_Data_Manipulate.ipynb <==
### Header

## Load Folder Drive
import os, sys
from google.colab import drive
drive.mount('/content/drive/')

## Move Pointer File ke DATA_UM
%cd /content/drive/MyDrive/DATA_UM

## Declare Root Project File
root_file = "/content/drive/MyDrive/DATA_UM"

import pandas as pd
import json
import csv
import re
import string
import numpy as np

### Dataset Online

df_online1 = pd.read_excel('dataset_online/filter/MANIPULATED_usable1_predict_id_tag_41.xlsx')
df_online2 = pd.read_excel('dataset_online/filter/MANIPULATED_usable2_review_tokped_tag_73.xlsx')
df_online3 = pd.read_excel('dataset_online/filter/MANIPULATED_usable3_smsa_dataset_tag_12.xlsx')
df_online4 = pd.read_excel('dataset_online/filter/MANIPULATED_usable4_online_shop_tag_18.xlsx')
df_online5 = pd.read_excel('dataset_online/filter/MANIPULATED_usable5_dataset_sa_ml_makanan_bayi_tag_30.xlsx')
df_online6 = pd.read_excel('dataset_online/filter/MANIPULATED_usable5_dataset_sa_ml_makanan_hewan_tag_57.xlsx')
df_online7 = pd.read_excel('dataset_online/filter/MANIPULATED_usable5_dataset_sa_ml_makanan_sarapan_tag_33.xlsx')
df_online8 = pd.read_excel('dataset_online/filter/MANIPULATED_usable5_dataset_sa_ml_tag_135.xlsx')

df_online1 = df_online1[df_online1['use'] == 1]
df_online1 = df_online1[['manipulated', 'use']]

df_online2 = df_online2[df_online2['use'] == 1]
df_online2 = df_online2[['manipulated', 'use']]

df_online3 = df_online3[df_online3['use'] == 1]
df_online3 = df_online3[['manipulated', 'use']]

df_online4 = df_online4[df_online4['use'] == 1]
df_online4 = df_online4[['manipulated', 'use']]

df_online5 = df_online5[df_online5['use'] == 1]
df_online5 = df_online5[['manipulated', 'use']]

df_online6 = df_online6[df_online6['use'] == 1]
df_online6 = df_online6[['manipulated', 'use']]

df_online7 = df_online7[df_online7['use'] == 1]
df_online7 = df_online7[['manipulated', 'use']]

df_online8 = df_online8[df_online8['use'] == 1]
df_online8 = df_online8[['manipulated', 'use']]

df_online_all = pd.concat([df_online1, df_online2, df_online3, df_online4, df_online5, df_online6, df_online7, df_online8])
df_online_all.reset_index(drop=True, inplace=True)
df_online_all[['text']] = df_online_all[['manipulated']]
df_online_all = df_online_all[['text']]
df_online_all

def change_tag(text):
    text = text.replace("[HASHTAG]", "#ultramilkxstraykids")
    text = text.replace("[HASHTAG_BRAND]", "#ultramilk")
    text = text.replace("[OFF_USERNAME]", "@ultramym")
    text = text.replace("[BA_USERNAME]", "@realstraykids")
    text = text.replace("[BA]", "StrayKids")
    text = text.replace("[RASA1]", "full cream")
    text = text.replace("[RASA2]", "coklat")
    text = text.replace("[RASA3]", "strayberry")
    text = text.replace("[RASA4]", "mocca")
    text = text.replace("[RASA5]", "taro")
    text = text.replace("[RASA6]", "caramel")
    text = text.replace("[RASA7]", "full cream")
    text = text.replace("[RASA8]", "coklat")
    text = text.replace("[RASA9]", "strayberry")
    text = text.replace("[RASA10]", "mocca")
    text = text.replace("[BRAND]", "ultramilk")
    text = text.replace("[BRAND_LAWAN]", "indomilk")
    text = text.replace("[BRAND_LAWAN2]", "frisianflag")
    text = text.replace("[EVENT]", "ultramilkxstraykids")
    text = text.replace("[WADAH]", "kotak")
    text = text.replace("[PLESETAN_EVENT]", "ultramilkxplastik")
    text = text.replace("[RASA11]", "taro")
    text = text.replace("[HASHTAG_LAWAN]", "#indomilkxastro")
    text = text.replace("[WADAH2]", "botol")
    text = text.replace("[KELOMPOK]", "susu")
    text = text.replace("[BRAND_LAWAN3]", "cimory")
    text = text.replace("[EVENT_NAME]", "ultramilkxstraykids")
    text = text.replace("[OFF_LAWAN]", "@indomilk")
    text = text.replace("[PRODUK]", "Ultramilk 250ml")
    text = text.replace("[UNGKAPAN_NEG1", "basi")
    text = text.replace("[UNGKAPAN_NEG2", "asem")
    text = text.replace("[UNGKAPAN_NEG3", "pait")
    text = text.replace("[UNGKAPAN_NEG4", "ga enak")
    text = text.replace("[UNGKAPAN_NEG5", "terlalu manis")
    text = text.replace("[UNGKAPAN_NEG6", "busuk")
    text = text.replace("[UNGKAPAN_NEG7", "menggumpal")
    text = text.replace("[UNGKAPAN_NEG8", "tidak segar")
    text = text.replace("[UNGKAPAN_NEG9", "rasanya gak jelas")
    text = text.replace("[UNGKAPAN_NEG10", "jamuran")
    text = text.replace("[UNGKAPAN_NEG11", "susu basi")
    text = text.replace("[UNGKAPAN_NEG12", "pait paitnya")
    text = text.replace("[UNGKAPAN_NEG13", "masam")
    text = text.replace("[UNGKAPAN_NEG14", "amis")
    text = text.replace("[UNGKAPAN_NEG15", "kurang fresh")
    text = text.replace("[UNGKAPAN_NEG16", "bau apek")
    text = text.replace("[UNGKAPAN_NEG17", "bau tidak enak")
    text = text.replace("[UNGKAPAN_NEG18", "asin")

    return text

## Running
df_online_all['text'] = df_online_all['text'].apply(lambda x: change_tag(x))

## Output Data
df_online_all

df_online_all['date'] = '2020-01-01'
df_online_all['platform'] = 'Dataset Online'
df_online_all['brand'] = 'Dataset Online'
df_online_all['hashtag'] = 'Dataset Online'
df_online_all['name'] = 'Dataset Online'
df_online_all['username'] = 'Dataset Online'
df_online_all['profile_id'] = 'Dataset Online'
df_online_all['comments'] = 0
df_online_all['share_retweet'] = 0
df_online_all['bookmark'] = 0
df_online_all['quotes'] = 0
df_online_all['likes'] = 0
df_online_all['link'] = 'Dataset Online'
df_online_all['labelling'] = 0

df_online_all = df_online_all[['date','platform','brand','hashtag','name','username','profile_id','comments','share_retweet','bookmark','quotes','likes','link','text','labelling']]
df_online_all

### Dataset Mentionlytics

df_mentionlytics_straykids = pd.read_excel('mentionlytics/straykids_ultramilk_mentionlytics_predicted.xlsx')
df_mentionlytics_itzy = pd.read_excel('mentionlytics/itzy_ultramilk_mentionlytics_predicted.xlsx')

df_mentionlytics_straykids = df_mentionlytics_straykids[df_mentionlytics_straykids['true_sen'].notna()]
df_mentionlytics_straykids = df_mentionlytics_straykids[['text', 'true_sen']]

df_mentionlytics_itzy = df_mentionlytics_itzy[df_mentionlytics_itzy['true_sen'].notna()]
df_mentionlytics_itzy = df_mentionlytics_itzy[['text', 'true_sen']]

df_mentionlytics = pd.concat([df_mentionlytics_straykids, df_mentionlytics_itzy])
df_mentionlytics.reset_index(drop=True, inplace=True)
df_mentionlytics = df_mentionlytics.rename(columns={'true_sen': 'labelling'})

df_mentionlytics['date'] = '2020-01-01'
df_mentionlytics['platform'] = 'Mentionlytics'
df_mentionlytics['brand'] = 'Mentionlytics'
df_mentionlytics['hashtag'] = 'Mentionlytics'
df_mentionlytics['name'] = 'Mentionlytics'
df_mentionlytics['username'] = 'Mentionlytics'
df_mentionlytics['profile_id'] = 'Mentionlytics'
df_mentionlytics['comments'] = 0
df_mentionlytics['share_retweet'] = 0
df_mentionlytics['bookmark'] = 0
df_mentionlytics['quotes'] = 0
df_mentionlytics['likes'] = 0
df_mentionlytics['link'] = 'Mentionlytics'

df_mentionlytics = df_mentionlytics[['date','platform','brand','hashtag','name','username','profile_id','comments','share_retweet','bookmark','quotes','likes','link','text','labelling']]
df_mentionlytics

print(len(df_mentionlytics_straykids))
print(len(df_mentionlytics_straykids[df_mentionlytics_straykids['true_sen'] == 0]))
print(len(df_mentionlytics_straykids[df_mentionlytics_straykids['true_sen'] == 1]))
print(len(df_mentionlytics_straykids[df_mentionlytics_straykids['true_sen'] == 2]))

print(len(df_mentionlytics_itzy))
print(len(df_mentionlytics_itzy[df_mentionlytics_itzy['true_sen'] == 0]))
print(len(df_mentionlytics_itzy[df_mentionlytics_itzy['true_sen'] == 1]))
print(len(df_mentionlytics_itzy[df_mentionlytics_itzy['true_sen'] == 2]))

### Dataset Under 2022

df_under_2022 = pd.read_excel('mentionlytics/!manipulate_under2020_done.xlsx', index_col=0)

## Running
df_under_2022['text'] = df_under_2022['text'].apply(lambda x: change_tag(x))

## Output Data
df_under_2022

print(len(df_under_2022))
print(len(df_under_2022[df_under_2022['labelling'] == 0]))
print(len(df_under_2022[df_under_2022['labelling'] == 1]))
print(len(df_under_2022[df_under_2022['labelling'] == 2]))

### Data Manipulation Brand Lain

df_manipulation = pd.read_excel(open('data/label/negative_manipulation.xlsx', 'rb'),sheet_name='dataset')
df_manipulation = df_manipulation.loc[:, ~df_manipulation.columns.str.contains('^Unnamed')]
df_manipulation = df_manipulation[df_manipulation['brand'] != 'ultramilk']
df_manipulation.reset_index(drop=True, inplace=True)
df_manipulation['brand'] = "Manipulation"
df_manipulation['hashtag'] = "Manipulation"
df_manipulation

## Running
df_manipulation['text'] = df_manipulation['text'].apply(lambda x: change_tag(x))

print(len(df_manipulation))
print(len(df_manipulation[df_manipulation['labelling'] == 0]))
print(len(df_manipulation[df_manipulation['labelling'] == 1]))
print(len(df_manipulation[df_manipulation['labelling'] == 2]))

### Merge ALL Data

df_additional_data = pd.concat([df_online_all, df_mentionlytics, df_under_2022, df_manipulation])
df_additional_data.reset_index(drop=True, inplace=True)
print(len(df_additional_data))

print(len(df_additional_data))
print(len(df_additional_data[df_additional_data['labelling'] == 0]))
print(len(df_additional_data[df_additional_data['labelling'] == 1]))
print(len(df_additional_data[df_additional_data['labelling'] == 2]))

df_additional_data.to_excel('dataset_online/filter/all_additional_data.xlsx')
