==> UltraJaya/Ultramilk_Final/1b_Merge_Data_Scraping_Manipulated.ipynb <==
## Load Folder Drive
import os, sys
import os.path
from os import path
from google.colab import drive
drive.mount('/content/drive/')

## Move Pointer File ke DATA_UM
%cd /content/drive/MyDrive/DATA_UM

## Declare Root Project File
root_file = "/content/drive/MyDrive/DATA_UM"

import pandas as pd
import json
import csv

df_twitter = pd.read_excel(open('data/label/RTP_FINAL_TWITTER.xlsx', 'rb'),sheet_name='dataset')
df_twitter = df_twitter.loc[:, ~df_twitter.columns.str.contains('^Unnamed')]
df_twitter = df_twitter[df_twitter['brand'] == 'ultramilk']
df_twitter.reset_index(drop=True, inplace=True)
print(len(df_twitter))

df_tiktok = pd.read_excel(open('data/label/RTP_FINAL_TIKTOK.xlsx', 'rb'),sheet_name='dataset')
df_tiktok = df_tiktok.loc[:, ~df_tiktok.columns.str.contains('^Unnamed')]
df_tiktok = df_tiktok[df_tiktok['brand'] == 'ultramilk']
df_tiktok.reset_index(drop=True, inplace=True)
print(len(df_tiktok))

df_ig = pd.read_excel(open('data/label/NOT_FINAL_IG.xlsx', 'rb'),sheet_name='dataset')
df_ig = df_ig.loc[:, ~df_ig.columns.str.contains('^Unnamed')]
df_ig = df_ig[df_ig['brand'] == 'ultramilk']
df_ig = df_ig[df_ig['add'] != 'a']
df_ig = df_ig[df_ig['labelling'] != 'not labelled']
df_ig = df_ig.drop(columns=['lang', 'conf_score', 'add', 'id', 'comment type', 'post type', 'post owner'])
df_ig = df_ig.drop(df_ig.columns[0], axis=1)
df_ig.reset_index(drop=True, inplace=True)
print(len(df_ig))

df_additional_data = pd.read_excel('dataset_online/filter/all_additional_data.xlsx')
df_additional_data = df_additional_data.loc[:, ~df_additional_data.columns.str.contains('^Unnamed')]
print(len(df_additional_data))

df_clean = pd.concat([df_twitter, df_tiktok, df_ig, df_additional_data], ignore_index=True)
df_clean = df_clean.dropna(subset=['text'])
print(len(df_clean))
print(len(df_clean[df_clean['labelling'] == 0]))
print(len(df_clean[df_clean['labelling'] == 1]))
print(len(df_clean[df_clean['labelling'] == 2]))
print(len(df_clean[df_clean['labelling'] == 3]))
print(len(df_clean[df_clean['labelling'] == 4]))
df_clean.to_excel('data/label/all_merge_data_training.xlsx')
df_clean.to_excel('dataset_online/filter/all_merge_data_training.xlsx')
