==> UltraJaya/Ultramilk_Final/3_DataPrepare4Training.ipynb <==
## **Load Root Google Drive File dan Import Library**

## Load Folder Drive
import os, sys
from google.colab import drive
drive.mount('/content/drive/')

## Move Pointer File ke DATA_UM
%cd /content/drive/MyDrive/DATA_UM

## Declare Root Project File
root_file = "/content/drive/MyDrive/DATA_UM"

# Import Important Library
import pandas as pd
import os.path
from os import path

## **STRUCTURE CHANGES TO MATCH TRAINING DATA REQUIREMENT**

## Declare Hashtag Data
hashtag = "ultramilk_all_data"

# Load Data for Training
df = pd.read_excel('data/preprocess/'+hashtag+'_all.xlsx')

# Rename column previous index as id_tweet for reporting purposes
df.rename( columns={'Unnamed: 0':'id_tweet'}, inplace=True)

# Show Data
df

## **CREATE FOLDER AND EXPORT DATA**

# Check Path is Available if no Create Dir
if path.exists(root_file+'/data/training/'+hashtag) == False:
  os.makedirs(root_file+'/data/training/'+hashtag)
else :
  print("File Already Existed")

# Export Data
df.to_csv('data/training/'+hashtag+'/'+hashtag+'_all.csv', index=False)
